feat: upgrade debian bookworm + alpine 3.21

+++ topology/ecorp/infra/provision.sh
@@ -9,13 +9,17 @@ DEBIAN_FRONTEND=noninteractive apt-get install -y certbot python3-certbot-apache
 
 # Hacker's mail account hacker@isp-a.milxc
 useradd -m -s "/bin/bash" -p `mkpasswd --method=sha-512 admin` admin || true
-addgroup admin mail
+adduser admin mail
 #mkdir /home/hacker/mail
 #touch /home/hacker/mail/Drafts /home/hacker/mail/Queue /home/hacker/mail/Sent /home/hacker/mail/Trash
 
-# disable systemd-resolved which conflicts with nsd
-echo "DNSStubListener=no" >> /etc/systemd/resolved.conf
-systemctl stop systemd-resolved
+DEB_VERSION=`cat /etc/debian_version | cut -d'.' -f1`
+if [ $DEB_VERSION -eq "11" ] # DEB 11 aka Bullseye
+then
+	# disable systemd-resolved which conflicts with nsd
+	echo "DNSStubListener=no" >> /etc/systemd/resolved.conf
+	systemctl stop systemd-resolved
+fi
 
 # manage isp-a.milxc zone
 apt-get update
